++ - kamangir/bolt#746

+++ bluer_ugv/swallow/session/classical/screen/video_player.py
@@ -10,22 +10,32 @@ from bluer_ugv.logger import logger
 
 
 class VideoPlayer:
-    def __init__(self):
+    def __init__(
+        self,
+        dryrun: bool = False,
+    ):
         self.process: Optional[subprocess.Popen] = None
         self.current_file: Optional[str] = None
 
         self.paused = False
+        self.dryrun = dryrun
 
-        logger.info(f"{self.__class__.__name__} created.")
+        logger.info(
+            "{} created{}.".format(
+                self.__class__.__name__,
+                "dryrun mode" if dryrun else "",
+            )
+        )
 
-    def pause(self):
-        if self.process and self.process.poll() is None:
-            try:
-                self.process.stdin.write(b"p")
-                self.process.stdin.flush()
-            except Exception as e:
-                logger.error(f"failed to send pause command: {e}")
-                return False
+    def pause(self) -> bool:
+        if not self.dryrun:
+            if self.process and self.process.poll() is None:
+                try:
+                    self.process.stdin.write(b"p")
+                    self.process.stdin.flush()
+                except Exception as e:
+                    logger.error(f"failed to send pause command: {e}")
+                    return False
 
         logger.info(
             "{}.{}".format(
@@ -45,25 +55,26 @@ class VideoPlayer:
             logger.error(f"file not found: {filename}")
             return
 
-        # Start omxplayer fullscreen.
-        loop_flag = "--loop" if loop else ""
-        cmd = f"omxplayer -b {loop_flag} {shlex.quote(filename)}"
-        logger.info(f"running: {cmd}")
+        if not self.dryrun:
+            # start omxplayer fullscreen.
+            loop_flag = "--loop" if loop else ""
+            cmd = f"omxplayer -b {loop_flag} {shlex.quote(filename)}"
+            logger.info(f"running: {cmd}")
 
-        # Kill previous playback if running
-        self.stop()
+            # Kill previous playback if running
+            self.stop()
 
-        try:
-            self.process = subprocess.Popen(
-                shlex.split(cmd),
-                stdin=subprocess.PIPE,
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-        except Exception as e:
-            logger.error(f"failed to run omxplayer: {e}")
-            self.process = None
-            return False
+            try:
+                self.process = subprocess.Popen(
+                    shlex.split(cmd),
+                    stdin=subprocess.PIPE,
+                    stdout=subprocess.DEVNULL,
+                    stderr=subprocess.DEVNULL,
+                )
+            except Exception as e:
+                logger.error(f"failed to run omxplayer: {e}")
+                self.process = None
+                return False
 
         self.current_file = filename
 
@@ -77,7 +88,10 @@ class VideoPlayer:
 
         return True
 
-    def play_list(self, playlist: List[str]):
+    def play_list(
+        self,
+        playlist: List[str],
+    ):
         log_list(
             logger,
             "play list",
@@ -94,16 +108,17 @@ class VideoPlayer:
                     self.process.wait()
 
     def stop(self) -> bool:
-        if self.process and self.process.poll() is None:
-            try:
-                self.process.stdin.write(b"q")
-                self.process.stdin.flush()
-            except Exception as e:
-                logger.warning(e)
-                pass
+        if not self.dryrun:
+            if self.process and self.process.poll() is None:
+                try:
+                    self.process.stdin.write(b"q")
+                    self.process.stdin.flush()
+                except Exception as e:
+                    logger.warning(e)
+                    pass
 
-            time.sleep(0.3)
-            self.process.kill()
+                time.sleep(0.3)
+                self.process.kill()
 
         self.process = None
